Impl Chrome-specific version of `peer_discovery`
