+++ ergo-rest/Cargo.toml
@@ -27,6 +27,7 @@ futures = "0.3"
 thiserror = "1"
 derive_more = "0.99"
 proptest-derive = {version = "0.3.0", optional = true }
+rand = "0.8.4"
 serde = { version = "1.0", features = ["derive"]}
 serde_json = { version = "1.0", optional = true }
 url = "2.2"
@@ -45,6 +46,7 @@ js-sys = "0.3"
 wasm-bindgen = "0.2"
 # Depdencies for `reqwest`
 serde_json = "1.0"
+getrandom = { version = "0.2", features = ["js"] }
 
 [target.'cfg(target_arch = "wasm32")'.dependencies.web-sys]
 version = "0.3.25"
